Move community build projects to normal sources
